requirements.txt
# Digital Twin Workshop - Python Dependencies
# Install with: pip install -r requirements.txt

# Groq AI Integration
groq>=0.4.0

# Upstash Vector Database
upstash-vector>=0.1.0

# Environment Variables
python-dotenv>=1.0.0

# Optional: Rich console output
rich>=13.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
